+++ app/Http/Controllers/SeasonController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Season;
+use App\Models\SeasonClothes;
 use Illuminate\Http\Request;
 
 class SeasonController extends Controller
@@ -38,14 +39,18 @@ class SeasonController extends Controller
     public function store(Request $request)
     {
 
-        $season = new Season();
+        $seasonCheck = Season::where('season_name',$request->input('season'))->first();
 
-        $season->season_name = $request->input($this->cName);
+        if(!empty($seasonCheck)){
+            return redirect()->route($this->cName.'.create')->with('error',"Товар з таким ім'ям вже існує");
+        }
+        else{
+            $season = new Season();
+            $season->season_name = $request->input($this->cName);
+            $season->save();
 
-        $season->save();
-
-        return redirect()->route($this->cName.'.create')->with('success','Сезон Успішно Додано');
-
+            return redirect()->route($this->cName.'.create')->with('success','Сезон Успішно Додано');
+        }
     }
 
     /**
@@ -88,6 +93,12 @@ class SeasonController extends Controller
      */
     public function destroy(string $id)
     {
+
+        $seasonCheck = SeasonClothes::where('season_id',$id)->first();
+
+        if(!empty($seasonCheck)){
+            return redirect()->route('clothes.index',['id'=>$id,'mode'=>'season'])->with('error','Даний сезон вже використовується в товарах:');
+        }
 
         $season = Season::where('season_id', $id)->first();
         $season->delete();
